backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
databricks-sql-connector==2.9.2
pandas==2.2.2
